Move network to cmd/nerdctl/network
