+++ cmd/nerdctl/network/network.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package network
 
 import (
 	"github.com/spf13/cobra"
@@ -22,7 +22,7 @@ import (
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/helpers"
 )
 
-func newNetworkCommand() *cobra.Command {
+func NewNetworkCommand() *cobra.Command {
 	networkCommand := &cobra.Command{
 		Annotations:   map[string]string{helpers.Category: helpers.Management},
 		Use:           "network",
