+++ cmd/nerdctl/network/network_create.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package network
 
 import (
 	"fmt"
