+++ cmd/nerdctl/network/network_inspect.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package network
 
 import (
 	"github.com/spf13/cobra"
