+++ cmd/nerdctl/network/network_inspect_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package network
 
 import (
 	"runtime"
