+++ cmd/nerdctl/network/network_list_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package network
 
 import (
 	"errors"
